Adds new implementation of Encoder.
- Starts a new subproject in the "logger" directory. This directory will
contain the Cobalt 1.0 client-side logic. Eventually the older
"encoder" directory will be deleted. For now we still depend on some
classes in the "encoder" directory.

- In a follow-up CL we will introduce a class named Logger that will act
as the C++ client-side interface to Cobalt. In this CL we introduce a
class named Encoder that will be used by the implementation of Logger.

- Encoder plays a role similar to the older Encoder class in the encoder
directory in that it is the coordinator for invoking the encoding
algorithms, generating ObservationMetadata, generating a random ID
for Observations etc.

- But this new Encoder differs from the older Encoder class in that it
is not the outward-facing API on the client side. This is an internal
API. One obvious difference is that this new Encoder class does not do
much of the validation that the older Encoder class did. That is because
the Logger will have to do much of this validation and we don't want to
do the same validation twice.

- Also introduces helper classes ProjectContext, Project and MetricID
to hold the client-side configuration for a project as well as pointers
to subsets of that configuration.

- Also introduces ProtoUtils as a package of utilities for dealing with
protos. Currently it has only one function and one typedef.

- Also introduces a new Status class that is intended to be a single Status
to serve all APIs in client-side Cobalt C++ code. New status codes may
be added as necessary.

Change-Id: I839a30fddceb202e540f6613c4a8335be1458477