Adds new Logger implementation.
Logger is the C++-client-facing interface to Cobalt.
It corresponds pretty closely to the FIDL interface.
It will be used by the FIDL service to implement the
FIDL interface.

The client invokes one of the Log*() methods and the Logger does a couple
of things with the logged Event:
- Possibly it generates and saves an immediate Observation
- Possibly it adds the Event to the Local Aggregator. (We only introduce
a hook for this to be coded later by pesk@).

Also:

- Adds an Event proto message. Even though we have previously introduced
many things that refer to an "Event" object, we have thusfar not introduced
any actual representation of an "Event" object and we do so here.
So far this is used only internally in
the Logger implementation. But I have made it a top-level proto because
I imagine it may be useful in other places such as in the API for the
Local Aggregator and possibly as a unit of storage in the Event Store
if we ever have one.

-Changes two methods to const.

Change-Id: I8729d972e96ef52474a9f3739adbc15a25ed0368