Change output of the config parser to something more structured.
In CobaltConfig, the repeated MetricDefinition field is replaced by
a list of CustomerConfig each of which contains a list of ProjectConfig,
each of which contains a list of MetricDefinition.

We introduce ProjectConfigYaml which is a single config file. It is
identical to the previous CobaltConfig but is not used by anyone outside
the parser.

testing=CQ and new test written.

Change-Id: I331f3a51445f8bec569e3b6c4d36574f4aa05325