Adds new class ProjectConfigs.
ProjectConfigs wraps a CobaltConfig and offers convenient and efficient
methods for looking up a project by name or ID. It will not be needed on
the Fuchsia client because there the CobaltConfig we get has a single
project. But it will be used in the Linux test_app and on the server
and in any context where we need to work with a CobaltConfig containing
multiple projects.

Change-Id: Iaefa07393d18b6047b8ede9a8a81d4844601cf60