ShippingManager: Set last_send_status when sending to Clearcut.
In the older logic for sending to the V0.1 Shuffler we set a variable
called last_send_status whenever we send an Envelope to the Shuffler.
This variable is checked by the test application.

In the newer logic for sending to Clearcut we were not do that and
this CL fixes that.

Change-Id: I514455f94c590e9c8f521123fdfb28b2bb5b101f