Fixes test flake in test_app_test.
This test has been flaky in the Cobalt CI.
See https://luci-scheduler.appspot.com/jobs/fuchsia/cobalt-x64-linux

The problem was that we were invoking operator= (the copy assignment
operator) on TestApp in one of the tests. (RunSendAndQuit.)
This is problematic because
it causes the fields of TestApp to be assigned to in lexical order.
This is problematic because it causes the existing values in
those fields to be destroyed in lexical order.
This is problematic because TestApp expects its fields to be
destroyed in reverse lexical order as this is how they would be
destroyed by the TestApp destructor. What was going on was that
one of the fields of TestApp was a unique_ptr holding a
MemoryObservationStore and a later field of TestApp was a unique_ptr
holding a ShippingManager. The ShippingManager has a background thread
that uses the ObservationStore but the ShippingManager does
not own the ObservationStore: it expects its invoker to manage it.
So it is the responsibility of the TestApp to manage the lifetime
of the ObservationStore and the ShippingManager and ensure that
the ObservationStore is not destroyed before the ShippingManager
is destroyed. (The ShippingManager destructor will block until
the ShippingManager's background thread stops.) This all works
fine as long as the fields of the TestApp are destroyed via the
TestApp's destructor. But it breaks down if the TestApps
fields are destroyed via TestApp::operator=.

A more robust solution would remove TestApp::operator=.

Change-Id: Ie5f5dec075f4c8f86b74f55b80123b8200da84af