Adds a copy of the Linux test app.
We are building a new version of the Linux test app for Cobalt 1.0.
In this preliminary CL we check-in a copy of the older version
of the app. In the next CL we give the new version of the app.
This will make it easier to see the diff between the two versions.

Change-Id: Ib673b8ef5b4731a17ddd5c494cb1adcced9fe1b4