New version of Linux Cobalt test application for Cobalt 1.0
This new version is similar to the previous version but it
is designed to work with Cobalt 1.0 and the new Logger interface
rather than the older Encoder interface.

In this CL we have not yet completed the new version. It
only handles interactive mode and it only handles one Log
command, namely log event.

Also the new version of the test_app_test is not yet complete.

I will do more work on both the test_app and the test_app_test in
follow-up CLs.

Change-Id: Idec97960b5c6623e551342763385893804e772fe