Switch strategies for incorporting tink into the CMake build.
We are now using the simpler add_subdirectory() method rather than
the more elaborate ExternalProject_Add() method. Why?
(1) Because we can. Since we completely control the CMake files for
Tink we can make sure they are compatible with this method.
(2) Because Tink depends on ABSL and ABSL wants to be added using
add_subdirectory(). I tried to switch ABSL to ExternalProject_Add()
and I wasn't able to.

Change-Id: I8b9b9a8a059da98ca83f9af24dc4df3a7f8c413c