Move ObservationBatch and Envelope into their own .proto files.
This CL goes towards reconciling the fact that we need to use our
.proto files in both the Cobalt repo and in Google 3. In some cases
we need to have two copies of a .proto file and in other cases we can
share the .proto file.

- SystemProfile, ObservationMetadata and ObservationBatch have been moved out
of observation.proto and put into a  new observation_batch.proto. This is
because we need to keep a separate copy of observation_batch.proto in the
Google 3 logs directory and it is useful to isolate this.

- Envelope has been moved out of observation.proto and put into a new
envelope.proto. This is because Envelope referes to ObservationBatch and
we need copy envelope.proto using CopyBara and use a regular expression
to update the reference to observation_batch.proto and so I wanted to
isolate it.

- The deprecated |backend| field has been removed from ObservationMetadata.

- The deprecated |system_profile| field in Envelope has been removed.
For a while now SystemProfiles are passed in ObservationMetadata instead
of Envelope.

- encrypted_message.proto was updated to match the version in Google 3.

Change-Id: I3b03d3ff40a12e74392e2bbf0353afec021e00bc