Adds report_row.proto.
report_row.proto contains proto messages that represent an individual row
of a generated report. There is one message per type of report. In this
CL I have only started the file with SimpleOccurrenceCountReportRow
because that is the type of report or which I am currently working on
the ReportGenerator.

This proto is used only on the Cobalt server--not at all on the client.
Nevertheless we want to keep it in the open source Cobalt repo as a way of
documenting precisely what the rows of each report type contain.

This CL also contains:

- Replaced the description of a report row in report_definition.proto
with a reference to the new proto.

- Added an entry to CMake file to build the new proto. Even though the
proto is not used on the client we still want to verify that it builds.

Change-Id: Ib75ac7073880c6c4624a14b50e3fdce220372734