Adds RapporConfigHelper
The logic for deriving RAPPOR configuration parameters based on a
MetricDefinition and a ReportDefinition needs to be shared between
the client and the server.

In this CL we factor that logic out of encoder.cc and logger.cc
into a new utility class RapporConfigHelper. We also move the
contents of rappor_settings.h to there and add a test.

Also in this CL: We added the method MetricRef::FullyQualifiedName().
This is to facilitate building a debug string for
RaporConfigHelper.

Change-Id: I8cbe65d6aa0399d9da5d12361fff828a63617b80